Adjust fan search (#1356)
* Fix fan search issues re: missing fans

* Fix issues with user profile, adjust event fan search

+++ db/src/models/fans.rs
@@ -1,7 +1,7 @@
 use chrono::NaiveDateTime;
 use uuid::Uuid;
 
-#[derive(Serialize, Clone)]
+#[derive(Clone, Debug, PartialEq, Queryable, Serialize)]
 pub struct DisplayFan {
     pub user_id: Uuid,
     pub first_name: Option<String>,
@@ -10,7 +10,7 @@ pub struct DisplayFan {
     pub phone: Option<String>,
     pub thumb_profile_pic_url: Option<String>,
     pub organization_id: Uuid,
-    pub order_count: Option<u32>,
+    pub order_count: Option<i64>,
     pub created_at: NaiveDateTime,
     pub first_order_time: Option<NaiveDateTime>,
     pub last_order_time: Option<NaiveDateTime>,
